update logic by adding isAuthenticated middleware

+++ backend/controllers/list.ts
@@ -1,17 +1,22 @@
 import { Request, Response } from "express";
 import ctrlWrapper from "../helper";
-import { ListItem } from "../types";
+import { ListItem, RequestUser } from "../types";
 const pool = require("../lib/db.ts");
 
 const getAllList = async (req: Request, res: Response) => {
-  const data = await pool.query("SELECT * FROM shopping_list ORDER BY id ASC");
+  const { id } = (req as unknown as RequestUser).user;
+  const data = await pool.query(
+    "SELECT * FROM shopping_list WHERE owner_id=$1 ORDER BY id ASC ",
+    [id]
+  );
   return res.status(200).json(data.rows);
 };
 
 const getOneItem = async (req: Request, res: Response) => {
   const { id } = req.params;
+  const { id: userId } = (req as unknown as RequestUser).user;
   const data = await pool.query(`SELECT * FROM shopping_list
-      WHERE id=${id}
+      WHERE id=${id} AND owner_id=${userId}
       `);
   if (data.rows.length == 0) {
     return res.status(404).json({ message: "That item is not exist!" });
@@ -21,14 +26,15 @@ const getOneItem = async (req: Request, res: Response) => {
 
 const addTOList = async (req: Request<{}, {}, ListItem>, res: Response) => {
   const { nameOfGood, description, store } = req.body;
+  const { id } = (req as unknown as RequestUser).user;
 
   if (!nameOfGood || !store) {
     return res.status(400).json({ message: "Fill all necessary fields!" });
   }
 
   const data = await pool.query(
-    `INSERT INTO shopping_list (good, description, store) 
-    VALUES($1, $2, $3) RETURNING * `,
+    `INSERT INTO shopping_list (good, description, store, owner_id) 
+    VALUES($1, $2, $3, ${id}) RETURNING * `,
     [nameOfGood, description, store]
   );
   return res.status(201).json(data.rows[0]);
@@ -39,22 +45,24 @@ const updateListItem = async (
   res: Response
 ) => {
   const { id } = req.params;
+  const { id: userId } = (req as unknown as RequestUser).user;
   const { nameOfGood, description, store } = req.body;
   if (!nameOfGood || !store) {
     return res.status(400).json({ message: "Fill all necessary fields!" });
   }
 
   const data = await pool.query(
-    `UPDATE shopping_list SET good = $1, description = $2, store = $3 WHERE id = ${id} RETURNING * `,
-    [nameOfGood, description, store]
+    `UPDATE shopping_list SET good = $1, description = $2, store = $3 WHERE id = ${id} AND owner_id=$4 RETURNING * `,
+    [nameOfGood, description, store, userId]
   );
   return res.status(200).json(data.rows[0]);
 };
 
 const deleteListItem = async (req: Request, res: Response) => {
   const { id } = req.params;
+  const { id: userId } = (req as unknown as RequestUser).user;
   const data = await pool.query(
-    `DELETE FROM shopping_list WHERE id=${id} RETURNING *`
+    `DELETE FROM shopping_list WHERE id=${id} AND owner_id=${userId} RETURNING *`
   );
 
   if (!data.rows[0]) {
